add tokens for update&delete of users

+++ les3/controllers/user/deleteUser.js
@@ -3,6 +3,11 @@ const {userService} = require('../../service');
 module.exports = async (req, res) => {
     try {
         const {id} = req.params;
+        const {id: token_id} = req.user;
+
+        if (+id !== token_id) {
+            throw new Error('It is not your user')
+        }
 
         await userService.deleteUserService(id);
